requirements.txt
pillow==10.4.0
pandas==2.2.3
gspread==6.2.1
streamlit==1.47.0